sync: Add new option '--force-sync'
Add new option to force updating the repo when it has different remote
URL from the one specified in the manifest. Without setting the option,
the function will just return error when detecting the situation.

Also update the function syncRepos() to reutrn error when any one of
those jobs fails.

Signed-off-by: Ramax Lo <ramaxlo@gmail.com>